--- .github/workflows/pylint.yml
@@ -1,6 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-- name: GitHub Action for pylint
-  # You may pin to the exact commit or the version.
-  # uses: cclauss/GitHub-Action-for-pylint@8ef4d22e119fb1cdc0f58f2e95cb1f8d8b0d55e6
-  uses: cclauss/GitHub-Action-for-pylint@0.7.0
