--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/home/jedrzej/.cache/pypoetry/virtualenvs/simulation-8aTaexDu-py3.9/bin/python"
-}
