+++ battle/simulation/config.py
@@ -1,0 +1,12 @@
+import logging
+
+BASE_UNIT_HP = 100
+SPECIAL_UNIT_HP = 300
+BASE_UNIT_STRENGTH = 50
+SPECIAL_UNIT_STRENGTH = 30
+BASE_UNIT_MOVEMENT_POINTS = 1
+SPECIAL_UNIT_MOVEMENT_POINT = 1
+BOARD_X = 20
+BOARD_Y = 20
+FRACTION_NAMES = ["Lemon", "Green", "Blue", "Transparent"]
+LOGGING_LEVEL = logging.CRITICAL
