+++ battle/simulation/sim_context.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Posision:
+    x: int
+    y: int
