--- crc.json
@@ -1,140 +1,0 @@
-[
-  {
-    "name": "Simulation",
-    "superclasses": "",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Defines the initial settings",
-      "Runs simulation"
-    ],
-    "collaborators": [
-      "Army, Board"
-    ]
-  },
-  {
-    "name": "Statistics",
-    "superclasses": "",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Saving statistics after each iteration"
-    ],
-    "collaborators": [
-      "Simulation"
-    ]
-  },
-  {
-    "name": "Army",
-    "superclasses": "",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores information about the army",
-      "Counts the number of active units"
-    ],
-    "collaborators": [
-      "Simulation, Unit"
-    ]
-  },
-  {
-    "name": "Board",
-    "superclasses": "",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores the board",
-      "Checks if the field is free",
-      "Counts the fields belonging to each army"
-    ],
-    "collaborators": [
-      "Simulation, Field"
-    ]
-  },
-  {
-    "name": "Field",
-    "superclasses": "",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores field modifiers",
-      "Stores information about units on the field ",
-      "Stores information about belonging to the army"
-    ],
-    "collaborators": [
-      "Board"
-    ]
-  },
-  {
-    "name": "Unit",
-    "superclasses": "",
-    "subclasses": "BaseUnit, Unit_A, Unit_B, Unit_C, Unit_D",
-    "type": 2,
-    "responsibilities": [
-      "Defines the basic skills and characteristics of the units"
-    ],
-    "collaborators": [
-      ""
-    ]
-  },
-  {
-    "name": "BaseUnit",
-    "superclasses": "Unit",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores the characteristics and skills of the base unit "
-    ],
-    "collaborators": [
-      "Army"
-    ]
-  },
-  {
-    "name": "SpecialUnitA",
-    "superclasses": "Unit",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores the characteristics and skills of special unit A "
-    ],
-    "collaborators": [
-      "Army"
-    ]
-  },
-  {
-    "name": "SpecialUnitB",
-    "superclasses": "Unit",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores the characteristics and skills of special unit B"
-    ],
-    "collaborators": [
-      "Army"
-    ]
-  },
-  {
-    "name": "SpecialUnitC",
-    "superclasses": "Unit",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores the characteristics and skills of special unit C"
-    ],
-    "collaborators": [
-      "Army"
-    ]
-  },
-  {
-    "name": "SpecialUnitD",
-    "superclasses": "Unit",
-    "subclasses": "",
-    "type": 1,
-    "responsibilities": [
-      "Stores the characteristics and skills of special unit D"
-    ],
-    "collaborators": [
-      "Army"
-    ]
-  }
-]
