+++ pyproject.toml
@@ -1,16 +1,20 @@
 [tool.poetry]
-name = "simulation"
-version = "1.0.2"
+name = "battle"
+version = "0.1.0"
 description = ""
-authors = ["Your Name <you@example.com>"]
+authors = ["Jędrzej Stasik <sjedrzej58@gmail.com>"]
 
 [tool.poetry.dependencies]
 python = "^3.9"
-mypy = "^0.812"
 pygame = "^2.0.1"
 
-
 [tool.poetry.dev-dependencies]
+pytest = "^6.2.4"
+pylint = "^2.8.3"
+radon = "^4.5.2"
+xenon = "^0.7.3"
+mypy = "^0.812"
+black = "^21.5b2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
