--- scripts/prepare_environment.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -Eeuo pipefail
-
-cd "$(dirname "$(readlink -f "$0")")"/..
-
-poetry install
