--- sim_context.py
@@ -1,6 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Posision:
-    x: int
-    y: int
