--- stats.py
@@ -1,11 +1,0 @@
-class Stast:
-    def __init__(self, BaseUnits: int, special_units: int, fiealds: int):
-        self.BaseUnits = BaseUnits
-        self.special_units = special_units
-        self.fiealds = fiealds
-
-    def save_to_json():
-        pass
-
-    def save_to_csv():
-        pass
